+++ Whim.sln
@@ -9,8 +9,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Whim.Dashboard", "src\Whim.
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Whim.WpfApp", "src\Whim.WpfApp\Whim.WpfApp.csproj", "{42B300E2-06B3-4671-A11F-25875BAB658C}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Whim.Bar", "src\Whim.Bar\Whim.Bar.csproj", "{E95331E9-FE75-44FF-ABDD-519550814DB6}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Whim.Gaps", "src\Whim.Gaps\Whim.Gaps.csproj", "{6B13ECFB-0754-4721-819A-9F4712B75E18}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Whim.Tests", "src\Whim.Tests\Whim.Tests.csproj", "{B2BA068D-D77A-4B7B-9B64-2C23701F48B5}"
@@ -23,7 +21,7 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Whim.FloatingLayout", "src\
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Whim.App", "src\Whim.App\Whim.App.csproj", "{D74F1389-F4D1-4A1C-ACE3-D91D70472DBB}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Whim.NewBar", "src\Whim.NewBar\Whim.NewBar.csproj", "{15DF5867-1318-4D25-A05D-0011F7C3D30D}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Whim.Bar", "src\Whim.Bar\Whim.Bar.csproj", "{D1CD9632-7BF6-40D3-BEE5-E6FA9BCE79CB}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -85,22 +83,6 @@ Global
 		{42B300E2-06B3-4671-A11F-25875BAB658C}.Release|x64.Build.0 = Release|Any CPU
 		{42B300E2-06B3-4671-A11F-25875BAB658C}.Release|x86.ActiveCfg = Release|Any CPU
 		{42B300E2-06B3-4671-A11F-25875BAB658C}.Release|x86.Build.0 = Release|Any CPU
-		{E95331E9-FE75-44FF-ABDD-519550814DB6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{E95331E9-FE75-44FF-ABDD-519550814DB6}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{E95331E9-FE75-44FF-ABDD-519550814DB6}.Debug|arm64.ActiveCfg = Debug|Any CPU
-		{E95331E9-FE75-44FF-ABDD-519550814DB6}.Debug|arm64.Build.0 = Debug|Any CPU
-		{E95331E9-FE75-44FF-ABDD-519550814DB6}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{E95331E9-FE75-44FF-ABDD-519550814DB6}.Debug|x64.Build.0 = Debug|Any CPU
-		{E95331E9-FE75-44FF-ABDD-519550814DB6}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{E95331E9-FE75-44FF-ABDD-519550814DB6}.Debug|x86.Build.0 = Debug|Any CPU
-		{E95331E9-FE75-44FF-ABDD-519550814DB6}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{E95331E9-FE75-44FF-ABDD-519550814DB6}.Release|Any CPU.Build.0 = Release|Any CPU
-		{E95331E9-FE75-44FF-ABDD-519550814DB6}.Release|arm64.ActiveCfg = Release|Any CPU
-		{E95331E9-FE75-44FF-ABDD-519550814DB6}.Release|arm64.Build.0 = Release|Any CPU
-		{E95331E9-FE75-44FF-ABDD-519550814DB6}.Release|x64.ActiveCfg = Release|Any CPU
-		{E95331E9-FE75-44FF-ABDD-519550814DB6}.Release|x64.Build.0 = Release|Any CPU
-		{E95331E9-FE75-44FF-ABDD-519550814DB6}.Release|x86.ActiveCfg = Release|Any CPU
-		{E95331E9-FE75-44FF-ABDD-519550814DB6}.Release|x86.Build.0 = Release|Any CPU
 		{6B13ECFB-0754-4721-819A-9F4712B75E18}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{6B13ECFB-0754-4721-819A-9F4712B75E18}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{6B13ECFB-0754-4721-819A-9F4712B75E18}.Debug|arm64.ActiveCfg = Debug|Any CPU
@@ -205,22 +187,22 @@ Global
 		{D74F1389-F4D1-4A1C-ACE3-D91D70472DBB}.Release|x86.ActiveCfg = Release|x86
 		{D74F1389-F4D1-4A1C-ACE3-D91D70472DBB}.Release|x86.Build.0 = Release|x86
 		{D74F1389-F4D1-4A1C-ACE3-D91D70472DBB}.Release|x86.Deploy.0 = Release|x86
-		{15DF5867-1318-4D25-A05D-0011F7C3D30D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{15DF5867-1318-4D25-A05D-0011F7C3D30D}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{15DF5867-1318-4D25-A05D-0011F7C3D30D}.Debug|arm64.ActiveCfg = Debug|Any CPU
-		{15DF5867-1318-4D25-A05D-0011F7C3D30D}.Debug|arm64.Build.0 = Debug|Any CPU
-		{15DF5867-1318-4D25-A05D-0011F7C3D30D}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{15DF5867-1318-4D25-A05D-0011F7C3D30D}.Debug|x64.Build.0 = Debug|Any CPU
-		{15DF5867-1318-4D25-A05D-0011F7C3D30D}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{15DF5867-1318-4D25-A05D-0011F7C3D30D}.Debug|x86.Build.0 = Debug|Any CPU
-		{15DF5867-1318-4D25-A05D-0011F7C3D30D}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{15DF5867-1318-4D25-A05D-0011F7C3D30D}.Release|Any CPU.Build.0 = Release|Any CPU
-		{15DF5867-1318-4D25-A05D-0011F7C3D30D}.Release|arm64.ActiveCfg = Release|Any CPU
-		{15DF5867-1318-4D25-A05D-0011F7C3D30D}.Release|arm64.Build.0 = Release|Any CPU
-		{15DF5867-1318-4D25-A05D-0011F7C3D30D}.Release|x64.ActiveCfg = Release|Any CPU
-		{15DF5867-1318-4D25-A05D-0011F7C3D30D}.Release|x64.Build.0 = Release|Any CPU
-		{15DF5867-1318-4D25-A05D-0011F7C3D30D}.Release|x86.ActiveCfg = Release|Any CPU
-		{15DF5867-1318-4D25-A05D-0011F7C3D30D}.Release|x86.Build.0 = Release|Any CPU
+		{D1CD9632-7BF6-40D3-BEE5-E6FA9BCE79CB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D1CD9632-7BF6-40D3-BEE5-E6FA9BCE79CB}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D1CD9632-7BF6-40D3-BEE5-E6FA9BCE79CB}.Debug|arm64.ActiveCfg = Debug|Any CPU
+		{D1CD9632-7BF6-40D3-BEE5-E6FA9BCE79CB}.Debug|arm64.Build.0 = Debug|Any CPU
+		{D1CD9632-7BF6-40D3-BEE5-E6FA9BCE79CB}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{D1CD9632-7BF6-40D3-BEE5-E6FA9BCE79CB}.Debug|x64.Build.0 = Debug|Any CPU
+		{D1CD9632-7BF6-40D3-BEE5-E6FA9BCE79CB}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{D1CD9632-7BF6-40D3-BEE5-E6FA9BCE79CB}.Debug|x86.Build.0 = Debug|Any CPU
+		{D1CD9632-7BF6-40D3-BEE5-E6FA9BCE79CB}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D1CD9632-7BF6-40D3-BEE5-E6FA9BCE79CB}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D1CD9632-7BF6-40D3-BEE5-E6FA9BCE79CB}.Release|arm64.ActiveCfg = Release|Any CPU
+		{D1CD9632-7BF6-40D3-BEE5-E6FA9BCE79CB}.Release|arm64.Build.0 = Release|Any CPU
+		{D1CD9632-7BF6-40D3-BEE5-E6FA9BCE79CB}.Release|x64.ActiveCfg = Release|Any CPU
+		{D1CD9632-7BF6-40D3-BEE5-E6FA9BCE79CB}.Release|x64.Build.0 = Release|Any CPU
+		{D1CD9632-7BF6-40D3-BEE5-E6FA9BCE79CB}.Release|x86.ActiveCfg = Release|Any CPU
+		{D1CD9632-7BF6-40D3-BEE5-E6FA9BCE79CB}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
