+++ src/Whim.App/Config.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using Whim.NewBar;
+using Whim.Bar;
 using Whim.FloatingLayout;
 using Whim.Gaps;
 //using Whim.TreeLayout;
@@ -30,15 +30,15 @@ internal static class Config
 							 new ColumnLayoutEngine(),
 							 new ColumnLayoutEngine("Right to left", false)
 							 );
-							 //treeLayoutEngine);
+			//treeLayoutEngine);
 
 			configContext.WorkspaceManager.Add(workspace);
 		}
 
 		// Add bar
-		List<BarComponent> leftComponents = new() { /* WorkspaceWidget.CreateComponent() */ };
-		List<BarComponent> centerComponents = new() { /* FocusedWindowWidget.CreateComponent() */ };
-		List<BarComponent> rightComponents = new() { /* ActiveLayoutWidget.CreateComponent(), DateTimeWidget.CreateComponent() */ };
+		List<BarComponent> leftComponents = new() { TextWidget.CreateComponent("Hello, world!"), WorkspaceWidget.CreateComponent(), };
+		List<BarComponent> centerComponents = new() { FocusedWindowWidget.CreateComponent() };
+		List<BarComponent> rightComponents = new() { ActiveLayoutWidget.CreateComponent(), DateTimeWidget.CreateComponent() };
 
 		BarConfig barConfig = new(leftComponents, centerComponents, rightComponents);
 		BarPlugin barPlugin = new(configContext, barConfig);
