+++ src/Whim.Bar/ActiveLayout/ActiveLayoutWidget.xaml
@@ -1,15 +1,10 @@
 ﻿<UserControl x:Class="Whim.Bar.ActiveLayoutWidget"
-			 xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-			 xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-			 xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-			 xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-			 xmlns:local="clr-namespace:Whim.Bar"
-			 xmlns:wpfui="clr-namespace:WPFUI.Controls;assembly=WPFUI"
-			 mc:Ignorable="d"
-			 d:DataContext="{d:DesignInstance Type=local:ActiveLayoutWidgetViewModel}"
-			 d:DesignHeight="450" d:DesignWidth="800">
-	<wpfui:Button Content="{Binding Path=ActiveLayoutEngine}"
-				  Command="{Binding Path=NextLayoutEngineCommand}"
-				  CommandParameter="{Binding}"
-				  Appearance="Secondary"/>
+	xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+	xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+	xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+	xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+	d:DesignHeight="450"
+	d:DesignWidth="800"
+	mc:Ignorable="d">
+	<Button Command="{x:Bind Path=ViewModel.NextLayoutEngineCommand}" Content="{x:Bind Path=ViewModel.ActiveLayoutEngine, Mode=OneWay}" />
 </UserControl>
