+++ src/Whim.Bar/ActiveLayout/NextLayoutEngineCommand.cs
@@ -7,17 +7,12 @@ public class NextLayoutEngineCommand : System.Windows.Input.ICommand
 {
 	private readonly IConfigContext _configContext;
 	private readonly ActiveLayoutWidgetViewModel _viewModel;
+	public event EventHandler? CanExecuteChanged;
 
 	public NextLayoutEngineCommand(IConfigContext configContext, ActiveLayoutWidgetViewModel viewModel)
 	{
 		_configContext = configContext;
 		_viewModel = viewModel;
-	}
-
-	public event EventHandler? CanExecuteChanged
-	{
-		add => CommandManager.RequerySuggested += value;
-		remove => CommandManager.RequerySuggested -= value;
 	}
 
 	public bool CanExecute(object? parameter) => true;
