+++ src/Whim.Bar/BarPlugin.cs
@@ -1,4 +1,4 @@
-using System;
+﻿using System;
 using System.Collections.Generic;
 
 namespace Whim.Bar;
