+++ src/Whim.Bar/DateTime/DateTimeWidget.xaml.cs
@@ -1,17 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Windows;
-using System.Windows.Controls;
-using System.Windows.Data;
-using System.Windows.Documents;
-using System.Windows.Input;
-using System.Windows.Media;
-using System.Windows.Media.Imaging;
-using System.Windows.Navigation;
-using System.Windows.Shapes;
+﻿using Microsoft.UI.Xaml.Controls;
 
 namespace Whim.Bar;
 
@@ -25,8 +12,8 @@ public partial class DateTimeWidget : UserControl
 	public DateTimeWidget(int interval = 100, string format = "HH:mm:ss dd-MMM-yyyy")
 	{
 		ViewModel = new DateTimeWidgetViewModel(interval, format);
-		InitializeComponent();
 		DataContext = ViewModel;
+		InitializeComponent();
 	}
