+++ src/Whim.Bar/FocusedWindow/FocusedWindowWidget.xaml
@@ -1,11 +1,10 @@
 ﻿<UserControl x:Class="Whim.Bar.FocusedWindowWidget"
-			 xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-			 xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-			 xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-			 xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-			 xmlns:local="clr-namespace:Whim.Bar"
-			 mc:Ignorable="d"
-			 d:DataContext="{d:DesignInstance Type=local:FocusedWindowWidgetViewModel}"
-			 d:DesignHeight="450" d:DesignWidth="800">
-	<TextBlock Text="{Binding Path=Value}"/>
+	xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+	xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+	xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+	xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+	d:DesignHeight="450"
+	d:DesignWidth="800"
+	mc:Ignorable="d">
+	<TextBlock Text="{x:Bind Path=ViewModel.Value, Mode=OneWay}" />
 </UserControl>
