+++ src/Whim.Bar/FocusedWindow/FocusedWindowWidget.xaml.cs
@@ -1,17 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Windows;
-using System.Windows.Controls;
-using System.Windows.Data;
-using System.Windows.Documents;
-using System.Windows.Input;
-using System.Windows.Media;
-using System.Windows.Media.Imaging;
-using System.Windows.Navigation;
-using System.Windows.Shapes;
+﻿using Microsoft.UI.Xaml.Controls;
 
 namespace Whim.Bar;
 
@@ -26,7 +13,6 @@ public partial class FocusedWindowWidget : UserControl
 	{
 		InitializeComponent();
 		ViewModel = new FocusedWindowWidgetViewModel(configContext);
-		DataContext = ViewModel;
 	}
 
 	public static BarComponent CreateComponent()
