+++ src/Whim.Bar/Whim.Bar.csproj
@@ -1,17 +1,27 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
+﻿<Project Sdk="Microsoft.NET.Sdk">
 	<PropertyGroup>
 		<TargetFramework>net6.0-windows10.0.19041.0</TargetFramework>
+		<TargetPlatformMinVersion>10.0.17763.0</TargetPlatformMinVersion>
+		<RootNamespace>Whim.Bar</RootNamespace>
+		<RuntimeIdentifiers>win10-x86;win10-x64;win10-arm64</RuntimeIdentifiers>
+		<UseWinUI>true</UseWinUI>
 		<Nullable>enable</Nullable>
-		<UseWPF>true</UseWPF>
 	</PropertyGroup>
-
 	<ItemGroup>
-	  <PackageReference Include="WPF-UI" Version="1.2.2-prerelease116" />
+		<None Remove="BarWindow.xaml" />
 	</ItemGroup>
 
+	<ItemGroup>
+		<PackageReference Include="Microsoft.WindowsAppSDK" Version="1.0.0" />
+		<PackageReference Include="Microsoft.Windows.SDK.BuildTools" Version="10.0.22000.194" />
+	</ItemGroup>
 	<ItemGroup>
 		<ProjectReference Include="..\Whim\Whim.csproj" />
 	</ItemGroup>
 
+	<ItemGroup>
+		<Page Update="BarWindow.xaml">
+			<Generator>MSBuild:Compile</Generator>
+		</Page>
+	</ItemGroup>
 </Project>
