+++ src/Whim.Bar/Workspace/WorkspaceWidget.xaml
@@ -1,48 +1,24 @@
-﻿<UserControl x:Class="Whim.Bar.WorkspaceWidget"
-			 xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-			 xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-			 xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-			 xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-			 xmlns:local="clr-namespace:Whim.Bar"
-			 xmlns:wpfui="clr-namespace:WPFUI.Controls;assembly=WPFUI"
-			 d:DataContext="{d:DesignInstance Type=local:WorkspaceWidgetViewModel}"
-			 mc:Ignorable="d"
-			 d:DesignHeight="450" d:DesignWidth="800">
+﻿<UserControl x:Class="Whim.Bar.WorkspaceWidget" xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation" xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml" xmlns:d="http://schemas.microsoft.com/expression/blend/2008" xmlns:local="using:Whim.Bar"
+	xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+	d:DesignHeight="450"
+	d:DesignWidth="800"
+	mc:Ignorable="d">
 	<UserControl.Resources>
 		<ResourceDictionary>
-			<ResourceDictionary.MergedDictionaries>
-				<ResourceDictionary Source="pack://application:,,,/WPFUI;component/Styles/Theme/Light.xaml" />
-				<ResourceDictionary Source="pack://application:,,,/WPFUI;component/Styles/Theme/Dark.xaml" />
-				<ResourceDictionary Source="pack://application:,,,/WPFUI;component/Styles/WPFUI.xaml" />
-			</ResourceDictionary.MergedDictionaries>
+			<local:BoolNegateConverter x:Key="intToBool" />
 		</ResourceDictionary>
 	</UserControl.Resources>
 
-	<ItemsControl ItemsSource="{Binding Workspaces}">
+	<ItemsControl ItemsSource="{x:Bind ViewModel.Workspaces}">
 		<ItemsControl.ItemsPanel>
 			<ItemsPanelTemplate>
-				<StackPanel Orientation="Horizontal"></StackPanel>
+				<StackPanel Orientation="Horizontal" />
 			</ItemsPanelTemplate>
 		</ItemsControl.ItemsPanel>
 
 		<ItemsControl.ItemTemplate>
-			<DataTemplate>
-				<wpfui:Button Content="{Binding Name}"
-							  Command="{Binding SwitchWorkspaceCommand}"
-							  CommandParameter="{Binding}"
-							  Appearance="Secondary">
-					<wpfui:Button.Style>
-						<Style TargetType="wpfui:Button" BasedOn="{StaticResource {x:Type wpfui:Button}}">
-							<Style.Triggers>
-								<DataTrigger Binding="{Binding ActiveOnMonitor}" Value="True">
-									<Setter Property="Background" Value="{Binding SystemFillColorSuccess}"/>
-								</DataTrigger>
-							</Style.Triggers>
-
-							<Setter Property="Margin" Value="0 0 4 0"/>
-						</Style>
-					</wpfui:Button.Style>
-				</wpfui:Button>
+			<DataTemplate x:DataType="local:WorkspaceModel">
+				<Button Command="{x:Bind SwitchWorkspaceCommand}" Content="{x:Bind Name}" />
 			</DataTemplate>
 		</ItemsControl.ItemTemplate>
 	</ItemsControl>
