+++ src/Whim.Bar/Workspace/WorkspaceWidgetViewModel.cs
@@ -1,5 +1,4 @@
 using System;
-using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.ComponentModel;
 using System.Linq;
