--- src/Whim.NewBar/BarWindow.xaml
@@ -1,13 +1,0 @@
-﻿<Window
-	x:Class="Whim.NewBar.BarWindow"
-	xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-	xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-	xmlns:local="using:Whim.NewBar"
-	xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-	xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-	mc:Ignorable="d">
-
-	<Grid>
-
-	</Grid>
-</Window>
